refactor: split generator wizard architecture
