+++ app/modules/generator/tests/test_validators.py
@@ -7,7 +7,7 @@ step, both happy and sad paths.
 
 from werkzeug.datastructures import MultiDict
 
-from app.modules.generator.services import safe_float
+from app.modules.generator.wizard_persisters import safe_float
 from app.modules.generator.validators import (
     validate_step1_form,
     validate_step2_form,
